+++ src/components/structure/carousels/carousel/components/carousel-page.js
@@ -42,13 +42,13 @@ var CarouselPage = function CarouselPage(_ref) {
     // If there's not enough items in the last page of the Carousel
     if (index >= pageItems.length) {
       return /*#__PURE__*/_react.default.createElement("div", {
-        key: index,
+        key: "page_" + index,
         className: "k-Carousel__page__item"
       });
     }
 
     return /*#__PURE__*/_react.default.createElement("div", {
-      key: index,
+      key: "page_" + index,
       className: "k-Carousel__page__item"
     }, /*#__PURE__*/_react.default.cloneElement(pageItems[index], itemProps));
   }));