+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Utilities: Add `k-u-padding*` classes.
+- Feature: Utilities: Add `k-u-margin-minus*` classes.
+
 ## [8.4.0] - 2022-02-14
 
 Features: