+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: `Button`: Change font size on `regular` size and up.
+
 ## [8.4.0] - 2022-02-14
 
 Features: